mqtt-server/requirements.txt
paho-mqtt
AWSIoTPythonSDK